Hyphenate words adjacent to NBSP
Previously, a sequence of words separated by NO-BREAK SPACE (U+00A0)
were treated as one long non-hyphenatable word. Now each such word
is processed separately and can get hyphenated.

Also fix an incorrect assumption that it's safe to cache words
separated by NBSP since NBSP can be used as a base for a combining
character, in which case it should not be separated from the
character following it.

Change-Id: Ia4351d29098d1f8015f797088d0255ea3a4de9df
Fixes: 27250514
Test: Manual (words separated by NBSP now get hyphenated)
Test: adb shell /data/nativetest/minikin_tests/minikin_tests