Compute vertical extents needed for layout
Text laid out in multiple fonts may need more vertical space in
order to make sure it doesn't overlap lines above and below it. But
the information was previously not available to users of Minikin.

Now we allow a parameter in various layout methods that returns the
ascent, desent, and line gap based on the maximum values of the fonts
actually used in laying out the text.

Also, LineBreaker now keeps and returns ascents and descents per
line, which callers can use to only allow extra linespacing on lines
that have such characters.

Finally, font run calculations have been modified to not check
support for \n and \r in cmap tables, as they are not needed. This
reduces the number of font runs and avoids cases where a fallback
font mistakenly gets used and increases the requested ascent and
descent.

Bug: 28963299
Bug: 29063863
Bug: 32057121
Bug: 37756858
Test: adb shell /data/nativetest/minikin_tests/minikin_tests
Test: bit FrameworksCoreTests:android.text.
Test: bit CtsTextTestCases:*
Change-Id: I959017210f19ced71db05b127148203dcf9a2d42