Make combining characters sticky
Let's say we have a combining character sequence, where the base
character is supported only by font B, while the combining character
is supported by both fonts A and B, where A is a higher-score font.

Previously, Minikin broke the sequence, and displayed the base
character in font B and the combining mark in font A. Now, we
display the whole sequence in font B, which should be able to handle
shaping and positioning correctly.

Bug: 65020024
Bug: 62419886
Test: adb shell /data/nativetest/minikin_tests/minikin_tests
Change-Id: I622787d2602d7fe806ebdadb5dcea74e6fc6321e