Restart word breaker from the start of the run when locale changed
If locale is changed, the current word break point may not be valid
since that position was calculated based on previous locale.
Need to re-calculate with the new locale.
To achive this, need to reset the internal current position of the
word breaker to the start position of the run.

After this change, addStyleRun should be used for setting locale
instead.

Test: bit CtsWidgetTestCases:android.widget.cts.TextViewTest
Test: bit CtsTextTestCases:android.text.cts.DynamicLayoutTest
Test: bit CtsTextTestCases:android.text.cts.StaticLayoutTest
Test: minikin_tests
Bug: 65175766
Bug: 65024629

Change-Id: Ic9034c04c9e27ede613e7a49b4fbf4a7c1067c28