Improve algorithm for greedy line breaker
Previously, we had a worst case O(N^2) algorithm for greedy line
breaking since we only saved the best line break, and needed to
recalculate the best break when the line still didn't fit after a
line break. Now we use an O(N lg N) algorithm that saves all the
best breaks in a queue in case we need to use some lesser candidates
later.

Bug: 63938206
Bug: 34185255
Test: adb shell /data/nativetest/minikin_tests/minikin_tests
Test: bit FrameworksCoreTests:android.text.
Test: bit CtsTextTestCases:*
Change-Id: I72b826a5461065cfb8795e918ba5a95b26c57879