Introduce HyphenatorMap for managing hyphenator in Minikin
The motivation of this map is porting Hyphenator.java to Minikin.
Since our locale is fully expanded by ICU, we need to search with
dropping some components from locale. See the comments in
HyphenatorMap::lookup for more details.

This CL only introduces HyphenatorMap. Nobody uses this class at this
moment. Hyphenator.java will be replaced with this class in subsequent
CLs.

Bug: 65024629
Test: minikin_tests

Change-Id: Ib6ff6101a011a9d1099b9d4e371fab0542c161a0