Separate greedy and optimal line breaking algorithms
1) The greedy algorithm and the optimal algorithm have been separated.
The greedy algorithm is now only run when needed (if there are tabs
in text, or a greedy algorithm has been asked for), as opposed to
every time.

2) The previous desperate line breaking assumed that the words are
supposed to end up in the same line as they would in the greedy
algorithm. This was wrong for optimal line breaking and resulted in
incorrect line breaks. We now limit desperate line breaking to the
greedy algorithm, and don't even add desperate candidates to the line
breaking candidates.

3) Extent calculations have been moved to when breaks are added as
opposed to during addition of style runs.  The running time is the
same, but the code is now simpler.

4) The fields, types, and methods in LineBreaker.h are reorganized so
they could be understood better.

Test: adb shell /data/nativetest/minikin_tests/minikin_tests
Test: bit FrameworksCoreTests:android.text.
Test: bit CtsTextTestCases:*
Bug: 65024629
Bug: 63938206
Bug: 28963299
Change-Id: If5e3481be6eba7e33f44c66151eb0f566ccee585
Fixes: 67018920