Remove mutex lock from HyphenatorMap constructor.
It is not a good idea to acquire a mutex lock in HyphenatorMap
constructor since HyphenatorMap is a singleton.
Usually HyphenatorMap is initialized in Zygote, so it should not
be a problem but it is problematic in unit tests.

Fortunately, unsupported language is very rare and it doesn't match
anything, so it just goes through the slow path in the lookup method
once, and gets added to the map at the end of the first call.

Bug: 65024629
Test: minikin_tests
Change-Id: I871bf87171d2c5351feddb294e46f97f354bf189