Move FontStyle to MinikinPaint member
FontStyle is always used with MinikinPaint except for itemization.
Also it is populated at the same time from Paint.
It would be good for it to be a member of MinikinPaint.

On the other hand, LocaleListId was stored in FontStyle is now used
everywhere, e.g. Hyphenator, WordBreaker, etc. It would be good
for it to be outside of FontStyle.

Bug: 65024629
Test: minikin_tests
Change-Id: I5199638ed21b0eaf1f426834933f64b603b6302a